Added code to verify existing downloaded models, before compression to allow mdoeals downloaded manually

+++ tasks/compress_file_task.py
@@ -19,9 +19,8 @@ class CompressFileTask(BaseTask):
         try:
 
             if os.path.exists(self.input_path_and_file_name) and os.path.exists(self.output_path_and_file_name):
-                self.logger.debug("Both input file and output file exist, assuming partially complete compression, removing %s and starting again.", self.output_path_and_file_name)
+                self.logger.debug("Both input file and output file exist, assuming interrupted partial compression, removing %s and starting again.", self.output_path_and_file_name)
                 os.remove(self.input_path_and_file_name)
-
 
             with tqdm(desc="Compressing Download", total=1, unit='B', unit_scale=True, leave=False, colour='blue') as progress_bar:
                 with py7zr.SevenZipFile(self.output_path_and_file_name, 'w', filters=[{"id": py7zr.FILTER_LZMA2, "preset": 8}]) as archive:
@@ -40,6 +39,7 @@ class CompressFileTask(BaseTask):
             return True
 
         except (Exception) as e:
-            self.logger.debug("Compressing failed, file removed partial compression %s", self.output_path_and_file_name)
-            os.remove(self.output_path_and_file_name)
+            self.logger.debug("Compressing failed, removed partially compressed file: %s", self.output_path_and_file_name)
+            if os.path.exists(self.output_path_and_file_name):
+                os.remove(self.output_path_and_file_name)
             return False
